// src/types/types.ts
export type NavigationLink = {
    href: string;
    label: string;
  };
  
  export type SocialLink = {
    href: string;
    icon: JSX.Element;
  };
  
  export type CarFinancingFeature = {
    id: number;
    message: string;
  };
  
  export type DealershipManagementData = {
    id: number;
    header: string;
    message: string;
  };
  
  export type QuestionData = {
    id: string;
    question: string;
    answer: JSX.Element;
  };
  
  export type WhyFeature = {
    title: string;
    description: string;
    image: string;
  };
  